refactor[frontend]: rename 'template' directory into 'base'
