+++ frontend/src/components/base/BaseFooter.vue
