+++ frontend/src/components/base/BaseNavbar.vue
